+++ src/sas/sascalc/dataloader/readers/ascii_reader.py
@@ -14,7 +14,7 @@
 
 import os
 import logging
-from sas.sascalc.dataloader.file_reader_base_class import FileReader
+from sas.sascalc.dataloader.filereader import FileReader
 from sas.sascalc.dataloader.data_info import DataInfo, plottable_1D
 from sas.sascalc.dataloader.loader_exceptions import FileContentsException,\
     DefaultReaderException
