+++ src/sas/sascalc/dataloader/readers/cansas_reader_HDF5.py
@@ -13,7 +13,7 @@ from sas.sascalc.dataloader.data_info import plottable_1D, plottable_2D,\
     Data1D, Data2D, DataInfo, Process, Aperture, Collimation, \
     TransmissionSpectrum, Detector
 from sas.sascalc.dataloader.loader_exceptions import FileContentsException, DefaultReaderException
-from sas.sascalc.dataloader.file_reader_base_class import FileReader, decode
+from sas.sascalc.dataloader.filereader import FileReader, decode
 
 from typing import Iterable
