+++ src/sas/sascalc/dataloader/readers/red2d_reader.py
@@ -16,7 +16,7 @@ import numpy as np
 from sas.sascalc.data_util.nxsunit import Converter
 
 from ..data_info import plottable_2D, DataInfo, Detector
-from ..file_reader_base_class import FileReader
+from ..filereader import FileReader
 from ..loader_exceptions import FileContentsException
